ft_ping: broadcast option
This commit adds support for broadcast messages with the
-b option. The user now can send broadcast messages to the
router that will be received by all hosts in the network and
depending on their firewall rules the pinger will receive responses.
no visual output change yet, but duplicates count needs to be introduced
and increment if received should only be done once and not on dups.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>