ft_traceroute: rafactor to select
use select instead of iterative loop that blocks, with timeout.
i still need to refactor some stuff to print duration of each probe
for time exceeded messages and print starts correctly if the probe
failed and we didnt get any response or because the probe got dropped
for some reason.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>